requirements.txt
aiohttp>=3.8.0
requests>=2.25.0
typing_extensions>=4.0.0
colorama>=0.4.4
dnspython>=2.2.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
tqdm>=4.64.0
aiodns>=3.0.0 